Added the useEffect() logic of the favicons and logo images in the component
